- ARM - problem with spurious crashes seems to be fixed by turning of FIQs, thanx to SII! - ARM interrupt processing - test code to check registers overflow. - ARM double sem test removed - can't be run twice! - vm thread resleep fixed (untested) - minor cleanups - EFI bios binary - slowly getting ready to AMD64 :)

+++ include/unix/uufile.h
@@ -40,7 +40,7 @@ struct uufileops
     errno_t     (*unlink)( struct uufile *f );
 
     //! rich man's ioctl - get property
-    errno_t	(*getproperty)( struct uufile *f, const char *pName, char *pValue, int vlen );
+    errno_t	(*getproperty)( struct uufile *f, const char *pName, char *pValue, size_t vlen );
     //! rich man's ioctl - set property
     errno_t	(*setproperty)( struct uufile *f, const char *pName, const char *pValue );
     //! rich man's ioctl - list properties
